+++ src/shapes/Star6.java
@@ -23,15 +23,39 @@ public class Star6 extends ShapeAbstract implements IShape {
     @Override
     public void drawFromXY(Graphics g, Color c, int x, int y, int width, int height, FillStatus fill) {
         g.setColor(c);
-        int[] xInts = new int[]{x, x + width / 3, x + width / 4, x + width / 2,
-                x + width - width / 4, x + width - width / 3, x + width,
-                x + width - width / 3, x + width - width / 4, x + width / 2,
-                x + width / 4, x + width / 3, x};
-        int[] yInts = new int[]{y + height / 2, y + height / 3, y, y + height / 4, y,
-                y + height / 3, y + height / 2, y + height / 3 * 2, y + height,
-                y + height / 4 * 3, y + height, y + height / 3 * 2, y + height / 2};
+        int[] xInts = new int[]{
+                x,
+                (int) (x + width / 3.0),
+                (int) (x + width / 4.0),
+                (int) (x + width / 2.0),
+                (int) (x + width - width / 4.0),
+                (int) (x + width - width / 3.0),
+                x + width,
+                (int) (x + width - width / 3.0),
+                (int) (x + width - width / 4.0),
+                (int) (x + width / 2.0),
+                (int) (x + width / 4.0),
+                (int) (x + width / 3.0),
+                x
+        };
+        int[] yInts = new int[]{
+                (int) (y + height / 2.0),
+                (int) (y + height / 3.0),
+                y,
+                (int) (y + height / 4.0),
+                y,
+                (int) (y + height / 3.0),
+                (int) (y + height / 2.0),
+                (int) (y + height / 3.0 * 2),
+                y + height,
+                (int) (y + height / 4.0 * 3),
+                y + height,
+                (int) (y + height / 3.0 * 2),
+                (int) (y + height / 2.0)
+        };
         if (fill == FillStatus.FULL) {
             g.fillPolygon(xInts, yInts, 13);
+        // TODO gradient fill
         } else if (fill == FillStatus.NONE) {
             g.drawPolygon(xInts, yInts, 13);
         }
